move cell back to utils, fix story
