+++ packages/grafana-ui/src/components/Table/DefaultCell.tsx
@@ -1,7 +1,6 @@
-import React, { FC, CSSProperties } from 'react';
+import React, { FC } from 'react';
 import { ReactTableCellProps } from './types';
 import { formattedValueToString } from '@grafana/data';
-import tinycolor from 'tinycolor2';
 
 export const DefaultCell: FC<ReactTableCellProps> = props => {
   const { field, cell, tableStyles } = props;
@@ -12,30 +11,4 @@ export const DefaultCell: FC<ReactTableCellProps> = props => {
 
   const displayValue = field.display(cell.value);
   return <div className={tableStyles.tableCell}>{formattedValueToString(displayValue)}</div>;
-};
-
-export const BackgroundColoredCell: FC<ReactTableCellProps> = props => {
-  const { cell, tableStyles, field } = props;
-
-  if (!field.display) {
-    return null;
-  }
-
-  const themeFactor = tableStyles.theme.isDark ? 1 : -0.7;
-  const displayValue = field.display(cell.value);
-
-  const bgColor2 = tinycolor(displayValue.color)
-    .darken(10 * themeFactor)
-    .spin(5)
-    .toRgbString();
-
-  const styles: CSSProperties = {
-    background: `linear-gradient(120deg, ${bgColor2}, ${displayValue.color})`,
-    borderRadius: '0px',
-    color: 'white',
-    height: tableStyles.cellHeight,
-    padding: tableStyles.cellPadding,
-  };
-
-  return <div style={styles}>{formattedValueToString(displayValue)}</div>;
 };
